Initial customization of theme, testing on github
This involved a lot of 'rudimentary' tweaks and customizations
that I won't bother listing here. Mostly things that are mentioned
in the 'theme setup' found here: https://mmistakes.github.io/minimal-mistakes/theme-setup/

For now we're lazy and leaving in all the 'how to do stuff' blog
posts, but those will be removed. In fact, this commit is mainly
testing that things work as expected on github