Tweak theme to include content from index.md in page about recent posts.
I thought it was a bit strange that the index.md file wasn't really used
for much in the theme; any text you put in it doesn't make it to the page.

This is probably an ugly, hacky way to have more than 'recent posts' on
the main page, but it works. The idea was to have a quick list of the next
events right on the homepage, so visitors won't have to hunt for the info.

This will probably need to be improved in the future to be much nicer