Remove posts that came with our theme, and use our readme instead of the theme's readme
This looks like the last step to having the site being 100%
cryptoparty, no more remnants of 'how to do this theme'.
publicly visible remnants anyways.

That said, if you want to refer to something for making posts,
feel free to look at older versions of this repo with 'git commit',
or perhaps even check out the main repo for the theme:
https://github.com/mmistakes/minimal-mistakes/